// src/role/enums/role.enum.ts
export enum RoleMessages {
    NOT_FOUND = 'Role not found',
    ALREADY_EXISTS = 'Role already exists',
    CREATED = 'Role session created successfully',
    UPDATED = 'Role sesion updated successfully',
    DELETED = 'Role deleted successfully',
    PRIVILEGE_ADDED = 'Privilege added successfully',
    PRIVILEGE_DELETED = 'Privilege deleted successfully',
    PRIVILEGE_ALREADY_EXISTS = 'Privilege already exists',
    PRIVILEGE_NOT_FOUND = 'Privilege not found',
}